pyproject.toml
[tool.poetry]
name = "preonboarding-boilerplate"
version = "0.1.0"
description = ""
authors = ["ShinJam <nevvjann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "^2021.3"
Django = "^3.2.8"
gunicorn = "^20.1.0"
newrelic = "^7.2.2"
psycopg2-binary = "^2.9.1"
dj-database-url = "^0.5.0"
django-model-utils = "^4.2.0"
django-unique-upload = "^0.2.1"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^21.1"
ipdb = "^0.13.9"
ipython = "^7.28.0"
mkdocs = "^1.2.3"
flake8 = "^4.0.1"
mock = "^4.0.3"
factory-boy = "^3.2.0"
coverage = "^6.0.2"
django-storages = "^1.12.2"
boto3 = "^1.19.3"
dj-rest-auth = "^2.1.11"
django-allauth = "^0.45.0"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"
pytest-mock = "^3.6.1"
pytest-responses = "^0.5.0"
pytest-sugar = "^0.9.4"
schema = "^0.7.4"
djongo = "^1.3.6"
Pillow = "^8.4.0"
django-configurations = "^2.2"
django-crontab = "^0.7.1"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.3"
pyclean = "*"
Werkzeug = "^2.0.2"
jupyter = "^1.0.0"
notebook = "^6.4.5"
requests = "^2.26.0"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
